move and rename folder errors, add docs, enum with note names and their indexes
